# [WIP] Use `cudaOccupancyMaxPotentialBlockSize` to determine kernel size