--- a/cupy/cuda/function.pxd
+++ b/cupy/cuda/function.pxd
@@ -11,6 +11,9 @@ cdef class Function:
         public Module module
         public intptr_t ptr
 
+    cpdef launch(
+        self, args, size_t gridx, size_t blockx, size_t shared_mem=*, stream=*)
+
     cpdef linear_launch(self, size_t size, args, size_t shared_mem=*,
                         size_t block_max_size=*, stream=*)