--- a/cupy/cuda/function.pyx
+++ b/cupy/cuda/function.pyx
@@ -17,6 +17,7 @@ from cupy.cuda cimport driver
 from cupy.cuda cimport runtime
 from cupy.cuda cimport stream as stream_module
 from cupy.cuda.texture cimport TextureObject
+from cupy import util
 
 
 cdef class CPointer:
@@ -165,6 +166,13 @@ cdef class Function:
             max(1, block[0]), max(1, block[1]), max(1, block[2]),
             args, shared_mem, s)
 
+    cpdef launch(
+            self, args, size_t gridx, size_t blockx, size_t shared_mem=0,
+            stream=None):
+        _launch(
+            self.ptr, gridx, 1, 1, blockx, 1, 1,
+            args, shared_mem, _get_stream(stream))
+
     cpdef linear_launch(self, size_t size, args, size_t shared_mem=0,
                         size_t block_max_size=128, stream=None):
         # TODO(beam2d): Tune it
@@ -172,8 +180,7 @@ cdef class Function:
             0x7fffffffUL, (size + block_max_size - 1) // block_max_size)
         cdef size_t blockx = min(block_max_size, size)
         s = _get_stream(stream)
-        _launch(self.ptr,
-                gridx, 1, 1, blockx, 1, 1, args, shared_mem, s)
+        self.launch(args, gridx, blockx, shared_mem, stream)
 
 
 cdef class Module: