@asi1024 2019-12-19
This PR is not so effective for the performance. (see https://github.com/cupy/cupy/pull/2827#issuecomment-566973948)