@leofang 2019-12-19
It could be that element wise operations are memory bound, so max occupancy would increase too many memory transactions and/or too much register pressure. Down-tune a bit the block size might help. Another potential factor to look at is to check data alignment.